Renamed package to jsonapiquery
