+++ circle.yml
@@ -9,6 +9,6 @@ test:
   pre:
      - pip install coveralls
   override:
-     - nosetests tests --with-coverage --cover-package=jsonapi_query
+     - nosetests tests --with-coverage --cover-package=jsonapiquery
   post:
      - coveralls
